+++ resources/assets/js_owl/admin/display/datatables.js
@@ -146,8 +146,7 @@ window.checkDateRange = (fromValue, toValue, value) => {
 window.columnFilters = {
     //date ========================================
     daterange (dateField, table, column, index, serverSide) {
-        let $dateField = $(dateField);
-
+        let $dateField = $(dateField)
         $dateField.on('apply.daterangepicker', function(e, date) {
             column.search($dateField.val());
         })
@@ -155,7 +154,6 @@ window.columnFilters = {
 
     //range ========================================
     range (container, table, column, index, serverSide) {
-        console.log('range')
         let $container = $(container),
             from = $('input:first', $container),
             to = $('input:last', $container),
@@ -209,8 +207,7 @@ window.columnFilters = {
 
     //select ========================================
     select (input, table, column, index, serverSide) {
-        let $input = $(input);
-
+        let $input = $(input)
         $input.on('change', () => {
             let selected = [];
             $input.find(':selected').each((i, e) => {