+++ resources/assets/js_owl/components/messages.js
@@ -3,10 +3,10 @@ module.exports = (function () {
     return {
 
         /**
-         * Вывод сообщения об ошибке
+         * Error message output
          *
-         * @param {String} title заголовок
-         * @param {String} message текст
+         * @param {String} title
+         * @param {String} message
          * @returns {*}
          */
         error(title, message) {
@@ -14,9 +14,10 @@ module.exports = (function () {
         },
 
         /**
-         * Вывод Success сообщения
-         * @param {String} title заголовок
-         * @param {String} message текст
+         * Success message output
+         *
+         * @param {String} title
+         * @param {String} message
          * @returns {*}
          */
         success(title, message) {
@@ -24,11 +25,11 @@ module.exports = (function () {
         },
 
         /**
-         * Вывод сообщения
+         * Message output
          *
-         * @param {String} title заголовок
-         * @param {String} message текст
-         * @param {String} type Тип сообщения (error, success)
+         * @param {String} title
+         * @param {String} message
+         * @param {String} type Message type (error, success)
          * @returns {*}
          */
         message(title, message, type) {
@@ -36,10 +37,10 @@ module.exports = (function () {
         },
 
         /**
-         * Вывод сообщения с подтверждением
+         * Confirmation message
          *
-         * @param {String} title заголовок
-         * @param {String} message текст
+         * @param {String} title
+         * @param {String} message
          * @param {Object} id
          */
         confirm(title, message, id) {
@@ -63,21 +64,25 @@ module.exports = (function () {
         },
 
         /**
-         * Вывод сообщения с полем ввода
+         * Displaying a message with an input field
          *
          * @param {String} title
          * @param {String} message
-         * @param {Function} callback Код выполняемый при подтверждении
-         * @param {String} inputPlaceholder Вспомогательный текст для поля ввода
+         * @param {Function} callback Code executed when confirmation
+         * @param {String} inputPlaceholder Placeholder
+         * @param {String} inputValue Default value in input field
+         * @param {String} imageUrl Show Image on this link
          */
-        prompt(title, message, inputPlaceholder) {
+        prompt(title, message, inputPlaceholder, inputValue, imageUrl) {
             return Swal.fire({
                 title: title,
                 text: message || '',
                 input: 'text',
                 showCancelButton: true,
-                closeOnConfirm: false,
                 inputPlaceholder: inputPlaceholder || '',
+                inputValue: inputValue || '',
+                imageUrl: imageUrl || '',
+                confirmButtonText: trans('lang.button.yes'),
                 cancelButtonText: trans('lang.button.cancel')
             })
         },