+++ resources/views/default/form/element/file.blade.php
@@ -10,7 +10,7 @@
 
 		@include(AdminTemplate::getViewPath('form.element.partials.helptext'))
 
-		<element-file @keyup.enter="return false;"
+		<element-file
 			url="{{ route('admin.form.element.file', [
 				'adminModel' => AdminSection::getModel($model)->getAlias(),
 				'field' => $path,
@@ -22,21 +22,23 @@
 			inline-template
 		>
 			<div>
-				<div v-if="errors.length" class="alert alert-warning">
+				<div v-if="errors.length" class="alert alert-warning" v-show="errors.length" style="display:none;">
 					<button type="button" class="close" data-dismiss="alert" aria-label="Close" @click="closeAlert()">
 						<span aria-hidden="true">&times;</span>
 					</button>
 
-					<p v-for="error in errors"><i class="fas fa-file-alt" aria-hidden="true"></i> @{{ error }}</p>
+					<p v-for="error in errors">
+						<i class="fas fa-file-alt" aria-hidden="true"></i> @{{ error }}
+					</p>
 				</div>
 
-				<div class="form-element-files clearfix" v-if="has_value">
+				<div class="form-element-files clearfix " v-if="has_value" v-show="has_value" style="display:none;">
 					<div class="form-element-files__item">
 						<div class="form-element-files__file">
 							<i class="fa-fw fas fa-file-alt"></i>
 						</div>
 						<div class="form-element-files__info">
-							<a :href="file" class="btn btn-default btn-xs pull-right">
+							<a :href="file" class="btn btn-default btn-xs pull-right" download title="{{ trans('sleeping_owl::lang.button.download') }}" target="_blank">
 								<i class="fas fa-cloud-upload-alt"></i>
 							</a>
 
@@ -49,7 +51,7 @@
 				</div>
 
 				<div v-if="!readonly">
-					<div class="btn btn-primary upload-button">
+					<div class="btn btn-primary upload-button btn-sm">
 						<i :class="uploadClass"></i> {{ trans('sleeping_owl::lang.file.browse') }}
 					</div>
 				</div>